requirements.txt
djangorestframwork